Rename package_fixtures/nested to package_fixtures/nested_modules
